+++ service/computeoptimizer/api.go
@@ -1113,11 +1113,21 @@ func (c *ComputeOptimizer) UpdateEnrollmentStatusRequest(input *UpdateEnrollment
 
 // UpdateEnrollmentStatus API operation for AWS Compute Optimizer.
 //
-// Updates the enrollment (opt in) status of an account to the AWS Compute Optimizer
-// service.
+// Updates the enrollment (opt in and opt out) status of an account to the AWS
+// Compute Optimizer service.
 //
 // If the account is a management account of an organization, this action can
 // also be used to enroll member accounts within the organization.
+//
+// You must have the appropriate permissions to opt in to Compute Optimizer,
+// to view its recommendations, and to opt out. For more information, see Controlling
+// access with AWS Identity and Access Management (https://docs.aws.amazon.com/compute-optimizer/ug/security-iam.html)
+// in the Compute Optimizer User Guide.
+//
+// When you opt in, Compute Optimizer automatically creates a Service-Linked
+// Role in your account to access its data. For more information, see Using
+// Service-Linked Roles for AWS Compute Optimizer (https://docs.aws.amazon.com/compute-optimizer/ug/using-service-linked-roles.html)
+// in the Compute Optimizer User Guide.
 //
 // Returns awserr.Error for service API and SDK errors. Use runtime type assertions
 // with awserr.Error's Code and Message methods to get detailed information about
@@ -3477,6 +3487,13 @@ type LambdaFunctionUtilizationMetric struct {
 	_ struct{} `type:"structure"`
 
 	// The name of the utilization metric.
+	//
+	// The following utilization metrics are available:
+	//
+	//    * Duration - The amount of time that your function code spends processing
+	//    an event.
+	//
+	//    * Memory - The amount of memory used per invocation.
 	Name *string `locationName:"name" type:"string" enum:"LambdaFunctionMetricName"`
 
 	// The statistic of the utilization metric.
@@ -4308,14 +4325,27 @@ func (s *ThrottlingException) RequestID() string {
 type UpdateEnrollmentStatusInput struct {
 	_ struct{} `type:"structure"`
 
-	// Indicates whether to enroll member accounts of the organization if the your
-	// account is the management account of an organization.
+	// Indicates whether to enroll member accounts of the organization if the account
+	// is the management account of an organization.
 	IncludeMemberAccounts *bool `locationName:"includeMemberAccounts" type:"boolean"`
 
 	// The new enrollment status of the account.
 	//
-	// Accepted options are Active or Inactive. You will get an error if Pending
-	// or Failed are specified.
+	// The following status options are available:
+	//
+	//    * Active - Opts in your account to the Compute Optimizer service. Compute
+	//    Optimizer begins analyzing the configuration and utilization metrics of
+	//    your AWS resources after you opt in. For more information, see Metrics
+	//    analyzed by AWS Compute Optimizer (https://docs.aws.amazon.com/compute-optimizer/ug/metrics.html)
+	//    in the Compute Optimizer User Guide.
+	//
+	//    * Inactive - Opts out your account from the Compute Optimizer service.
+	//    Your account's recommendations and related metrics data will be deleted
+	//    from Compute Optimizer after you opt out.
+	//
+	// The Pending and Failed options cannot be used to update the enrollment status
+	// of an account. They are returned in the response of a request to update the
+	// enrollment status of an account.
 	//
 	// Status is a required field
 	Status *string `locationName:"status" type:"string" required:"true" enum:"Status"`